fix: restore TC-UMES-006 transformFormData test, move SPEC-042/043 to TC-UMES-008
TC-UMES-006 was overwritten during rebase conflict resolution. This
restores the original transformFormData applyToForm opt-in test and
moves the SPEC-042 + SPEC-043 multi-ID / notification handler test
to TC-UMES-008.
